Move IO functions into separate files
