+++ io.c
@@ -1,0 +1,11 @@
+#include <stdio.h>
+#include "io.h"
+
+void printUsage(char* programName) {
+    printf("Usage:\n");
+    printf("%s search [--exact] <package>\n", programName);
+    printf("%s install [--exact] <package>\n", programName);
+    printf("%s upgrade\n", programName);
+    printf("%s clean\n", programName);
+}
+
