+++ io.h
@@ -1,0 +1,8 @@
+#ifndef IO_H_
+#define IO_H_
+
+static const char DIVIDER[] = "####################";
+
+void printUsage(char* programName);
+
+#endif
